refactor: rename session files and directories to conversation
