+++ ts/packages/shell/src/main/conversationManager.ts
